First attempt at gracefully recovering from network errors.
Retry the AudioStream and seek back to where we were before. This doesn't work perfectly because the seek operation doesn't work as soon as the stream is started, but rather it needs a bit of time to calculate the bit rate and such.

The delay of a half second seems a bit wonky at times, but it's better than nothing!