Don't re-fetch images twice
When a new song comes in, it's added to the history queue, so the artwork is
needed for both there and the main playback screen, so ImageLoader is queried
twice for the same image. This is obviously not ideal, but easy to fix by
satisfying all requests in the queue of images to load to the same url when
another finishes loading.