Clean up Hermes, View and Pandora menus.
Hermes menu: Fix capitalization of "Check for Updates".

View menu: Indent "History"/"Stations" under "Show/Hide Drawer" to make it clearer how the menu items behave.  Display checkmark next to "History"/"Stations" if visible, or mixed state indicator (hyphen) if not.

Pandora menu: Remove item to play current station as it's now possible to do so with the return/enter key (and this was stealing the Minimize keyboard shortcut in any case).  Move "Next" and rename to "Skip to Next Song" as this frequently confused me.