Add some remaining functionality to status menu.
"Tired of Song" is visible always; "Check for Updates" and "About Hermes" replace "Hermes Preferences" and "Quit Hermes" with the Option key down.

Everything else not directly appearing in the menu is usable when the menu bar is hidden by its keyboard equivalent. Toolbar editing/customization is also available by the title/toolbar context menu.

We may want to consider further enhancements to this menu in future. I don't actually use the menu bar-hidden mode myself, so it's hard to know.