Redo the Credits file. Remake as RTF.
Both the RTF version and the HTML version are not aligning well, but the
RTF version appears to be more centered. Added links to Homepage,
GitHub, Issues, and Changelog in About as well since users will not
have access to the Help menu if using the status bar item.

Added authors with links to Github pages.