Clean up station seed editor.
Show seed kinds as groups (and don’t make them selectable).

Make it clearer that the search box is a search box.

Don’t search for nothing (which gave a Pandora error).

Clear the search box when loading a station; clear the search results when the search box is cleared.

Only enable add/remove buttons when appropriate.

Sort alphabetically so kind ordering isn’t arbitrary.

Only include seed kinds which are in use.