Fix new station search results showing up in the station editor and vice versa.
This eliminates the Pandora object being the ‘object’ in notifications, which isn’t terribly useful as it’s a singleton, and isn’t actually used anywhere that I can find.

It frees up ‘object’ in notifications to be used in future for something else - such as a request identifier so multiple identical in-flight requests can be distinguished.  Currently I use it for a search string, which provides a very crude form of disambiguation between the two places in the app where you can search.