allowAddMusic lies.
Undoes 7247e41c932c73ab22f90b2729e77c463cc822ac, as well as some more (unused) code which should have been removed in that commit.

It seems that (with the exception of QuickMix, which is excluded from editing elsewhere) that this is not actually a limitation any more; it's possible to add seeds to genre stations as well (#267).