Status item fixes (#275).
- Remove extra space; it looks fine without it.
- Fix text baseline (why this isn’t the default…).
- Remove some old deprecated API usage.

Still an issue: no spacing on the left side of the album art when we truncate.

A simple solution might be to dump the fanciness and go back to ellipsizing ourselves (though should do it with string width, not character count).

It might be possible to re-add left padding with Auto Layout, but this involves depending on properties of undocumented views and I don’t want to go there.