Clean up proxy preferences handling.
Have a sane tab order in the nib.

Fix wording and some copy-paste errors with help tags.

Validate proxy hosts/ports.  Previously was crashing (!) if invalid.

Use a number formatter to validate the port.

For the host, we don’t do any validation of the field, but at least strip leading/trailing whitespace and display a message if it doesn’t work.

Have only tested a HTTP(S) proxy, not a SOCKS proxy, so there may still be some lurking issues.